Programmable Cards, Canceling Mechanic, Improved Player UX
- make player back to monster controller and have playerController on theri monstor for player componenet stuff
- UX Improvement: hovering over enemies colors monster and ticks / sorting
- Completed winning battle condition
- Implemented Attack canceling (odd bug complaining about index which needs to be fixed)
- Made cards programmable

+++ Assets/Scripts/Monster/MonsterBase.cs
@@ -9,9 +9,9 @@ public abstract class MonsterBase
     public string description;
     public string spriteFile;
     public int baseHealth;
-    public int baseAttack;
-    public int baseDefense;
-    public int baseSpeed;
+    public float baseAttack;
+    public float baseDefense;
+    public float baseSpeed;
     public BaseType baseType;
     public List<Attack> moveSet = new List<Attack>();    //defines what attacks this monster has
